back-office; sans modif produits, sous-cat, tailles et couleurs, mais avec redirection de page ecommerce, par fichier constants.

+++ includes/banque_retour.php
@@ -3,28 +3,37 @@
 
 <?php
 include '../bdd/bd.php';
+include 'constants.php';
 $trans = $_POST['transaction'];
-echo $trans;
-
-$reqCommandes = $bd->prepare('INSERT INTO commandes SELECT * FROM panier WHERE panier_transaction = :trans');
-$reqCommandes->bindValue(':trans', $trans, PDO::PARAM_INT);
-$reqCommandes->execute();
+//echo $trans;
 
 
-$reqPanierSupprime = $bd->prepare('DELETE FROM panier WHERE panier_transaction = :trans');
-$reqPanierSupprime->bindValue(':trans', $trans, PDO::PARAM_INT);
-$reqPanierSupprime->execute();
+if (isset($bd)) {
+    $reqCommandes = $bd->prepare('INSERT INTO commandes SELECT * FROM panier WHERE panier_transaction = :trans');
 
-unset($_SESSION['total']);
+    $reqCommandes->bindValue(':trans', $trans, PDO::PARAM_INT);
+    $reqCommandes->execute();
 
-$infoSuccess= '<div class="alert alert-success" role="alert">
- Votre transaction No. '.$trans.' s\'est déroulée avec succèes !
+
+    $reqPanierSupprime = $bd->prepare('DELETE FROM panier WHERE panier_transaction = :trans');
+    $reqPanierSupprime->bindValue(':trans', $trans, PDO::PARAM_INT);
+    $reqPanierSupprime->execute();
+
+
+    unset($_SESSION['total']);
+
+    $infoSuccess = '<div class="alert alert-success" role="alert">
+ Votre transaction No. ' . $trans . ' s\'est déroulée avec succèes !
 </div>';
 
-$_SESSION['infoPaiementSucces'] = $infoSuccess;
+    $_SESSION['infoPaiementSucces'] = $infoSuccess;
+    echo $_SESSION['infoPaiementSucces'];
+    echo URL . '/index.php?page=71';
 
-echo '<script> window.location.replace("http://localhost/Formation_cci/e_commerceV2/index.php?page=71");</script>';
+    header('Location:'.URL . '/index.php?page=71');
 
+}
 
-?>
+//echo '<script> window.location.replace('.URL.'"/index.php?page=71");</script>';
+
